code/backend/requirements.txt
fastapi
uvicorn
pypdf2
python-docx
sentence-transformers
faiss-cpu
transformers

extract-msg